move convert_polar<->cartesian to helpers
